implment optimisic ui in favorite
The favorite button was 'laggy' bc it was sending a request
to the server, getting a response, then updating the value.
Instead of having a delay, we are using a strategy called
'optimistic ui' which pre-fills in the value in the ui while
sending it to the server, then updates the value with the server
response.